Created New Headers and C Files, I think removed Magic Numbers, added beginning of LL
