+++ dependencies.h
@@ -19,133 +19,44 @@
 #include "rlgl.h" //required by raylib.h
 #include "typedefs.h" // included for function prototypes
 #include <ctype.h> // included for tolower
-
+#include "drawing.h" // included for functions that change the window
+#include "inits.h" // included for functions that run at startup
+#include "navigations.h" // inlcuded for functions that relate to the movement and navigation of AEDVs
+#include "linkedlist.h"
 #include <malloc.h> // included for dynamic allocation using malloc
 
+/* Magic Number Defines */
+    //directions
+    #define SOUTH 0
+    #define NORTH 1
+    #define EAST 2
+    #define WEST 3
+    //default settings
+    #define DEFAULTZOOM 0.94f
+    #define DEFAULTOFFSET 45
+    #define DEFAULTFONTSIZE 20
+    //others
+    #define EVINBASE 10000
 
-#define SOUTH 0
-#define NORTH 1
-#define EAST 2
-#define WEST 3
-
-extern int MAX_COLS, MAX_ROWS, cellWidth, cellHeight;
+/* Global Externs */
+extern int maxAEDV, MAX_COLS, MAX_ROWS, cellWidth, cellHeight;
 extern Tile** dynamicMap;
+extern AEDV *listOfVehicles[4];
+extern const int screenWidth;
+extern const int screenHeight;
+extern Camera2D camera;
+extern AEDV_Node ** ActiveList;
 
-/**
- * @brief Function draws the map and then calls the navigations function
- */
-void UpdateDrawFrame(void);     // Update and Draw one frame
-/**
- * @brief Draws the map of the city using the initalized Tile array
- * @param tile Struct containing Tile type information
- */
-void DrawMap(Tile tile);
-/**
- * @brief Calculates and then updates the position of the AEDV in the map
- * @param vehicle AEDV to move around
- * @note algorithm works on a two-way rule free road system
- */
-void MapNavigation(AEDV *vehicle);
-/**
- * @brief Calculates North/South Movement
- * @param vehicle AEDV to move
- * @param direction North/South
- */
-void AvenueMovement(AEDV *vehicle, int direction);
-/**
- * @brief Calculates East/West Movement
- * @param vehicle AEDV to move
- * @param direction Esat/West
- */
-void StreetMovement(AEDV *vehicle, int direction);
-/**
- * @brief allows for moving and resizing of the camera
- */
-void CameraControl();
-/**
- * @brief  the size of the map array using the input rows and columns
- */
-void allocDynamicMap();
+
 /**
  * @brief function takes user input and then calls InitAEDV to assign provided values
  */
-void AEDVInput();
-/**
- * @brief populates AEDV struct with inital values
- */
-void InitAEDV(AEDV *vehicle, int locationX, int locationY, int destinationX, int destinationY, int identifierCode);
-/**
- * @brief initializes values of the tiles in the map
- * @note Uses setPerimeterRoads() and setInternalTiles()
- */
-void InitTiles();
-/**
- * @brief sets the values of the boundary roads (either two way roads or junctions)
- * @param i col
- * @param j row
- */
-void setPerimeterRoads(int i, int j);
-/**
- * @brief sets the value of the internal tiles according to road directions set
- * @param i
- * @param j
- * @param firstStreetDirection
- * @param firstAvenueDirection
- */
-void setInternalTiles(int i, int j, int firstStreetDirection, int firstAvenueDirection);
-/**
- * @brief One way pathing navigation algorithm
- *
- * @param vehicle
- *
- * @note Algorithm is dumb, only checking one tile ahead and without considering rules of the road. Ommitted from Task 5 Code
- */
-void OneWayNavigation(AEDV * vehicle);
-/**
- * @brief function used for one way navigation, determines whether the input direction is able to be travelled
- * @param vehicle
- * @param direction
- * @note used in one way, omitted for task 5
- * @return True or False
- */
-bool IsValidDirection(AEDV* vehicle, int direction);
-/**
- *
- * @param currentTile
- * @param vehicle
- * @note used in one way, omitted for task 5
- */
-void StreetMoves(int currentTile, AEDV *vehicle);
-/**
- *
- * @param vehicle
- * @param destinationTile
- * @param desiredJunctions
- * @note used in one way omitted in task 5
- */
-void FindAdjacentJunctions(AEDV* vehicle, int destinationTile, Cord* desiredJunctions);
-/**
- *
- * @param vehicle
- * @param desiredJunctions
- * @param destinationTile
- * @note utilized in one way, omitted in task 5
- * @return
- */
-bool AtDesiredJunction(AEDV* vehicle, Cord* desiredJunctions, int destinationTile);
-/**
- * @brief uses generation rules to determine if the tile chosen is a deliverable location
- * @param col
- * @param row
- * @note function uses col and row instead of a tile type because it can be called before tiles have been generated
- * @return
- */
-bool isValidDestination(int col, int row);
 
-void SetupInitialConditions();
 
-void UpdateMap();
 
-void AssignNewOrders();
+
+
+
+
 
 #endif //EXAMPLE_DEPENDENCIES_H
